Time: 0 ms (100.00%) | Memory: 53.5 MB (80.63%) - LeetSync

+++ 528-swapping-nodes-in-a-linked-list/swapping-nodes-in-a-linked-list.c
@@ -9,7 +9,7 @@ struct ListNode* swapNodes(struct ListNode* head, int k) {
     if(head==0 || head->next==0)
     return head;
 struct ListNode* p=head,*q;
-long i,c=0,j;
+int i,c=0,j;
 while (p)
 {
     c++;
